requirements.txt
selenium
# playwright==1.49.0
python-dotenv==1.0.1
requests==2.32.3
pprintpp
